Added different Lint types, added Parsed_Command abstraction
The Parsed_Command abstraction groups the raw command with its
equivalent parsed structure, that is evaluated lazily. This allows
decoupling the lints from document structure. This reduces the overhead
of implementing easier lints that don't require a complicated structure.

The different lint types are the following:
- Structure_Lint: This is the old Lint type. It uses the pasred document
structre to generate the lints. Implicit_Rule is a lint that uses this
abstraction.
- Raw_Command_Lint: this uses the raw command to generate the lint.
Especially suitable for lints that detect unwanted commands (like
sorry). No_Sorry is a lint that uses this abstraction.
- Raw_Token_Stream_Lint: this uses the raw token stream
- Parser_Lint: this implements the lints as parser combinator, that
parses tokens into a Lint_Report. This is suitable for lints that don't
need a complicated structure. Short_Definition uses this abstraction