Introduced `Lint_Report` and `Single_Command_Lint`
The `Lint` trait was generalized to accept a list of parsed commands
instead of just one (which will be helpful in the future). The
`Lint_Report` is now just a wrapper, but it provides a helpful
abstraction so that lints can for example have more than one result. It
might be also used to provide a quick association between commands and
their lint results (for example, through a
`Map[Document.Command_Id, List[Lint_Result]])`